Worked on common data model
Switched names from Event to Command, because we want the server to
be allowed to drop commands when they are non-sensical. I think that
name fits better.

Also, I've implemented the changing mechanism for fields now. And I
have added information on which users are affected when competences
or evidences are added or removed.