Hacky, but limit size of side menu to fit screen
The layouting primitives do not work out perfectly, but my
understanding of them improves. Most important thing: when
something is shrinkable, it must have the MinH0 or MinW0
attribute, because otherwise it is set to its content-position.

Also HFull and WFull must be passed to the intermediate divs that
I keep adding to mount components. As a result, I probably want
to add layout primitives that enforce that something does not
overflow or make this a general paradigm.